RFCT: move uploadform to a subfeature folder and extract toast outside for readability
